Reorganize benchmarks
Splits apart the `Uint` and `DynResidue` benchmarks, in preparation for
adding benchmarks for `BoxedUint` and `BoxedResidue`.
